requirements.txt
# Urban Issue Reporter Dependencies
Flask==2.3.3
Werkzeug==2.3.7

# Database
pymongo==4.6.1
dnspython==2.4.2

# Environment and Configuration
python-dotenv==1.0.0

# ML Dependencies for Automation
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2

# Testing (Optional but recommended)
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0

# Production Server (Optional - for deployment)
# gunicorn==21.2.0
# waitress==2.1.2

# Rate Limiting (Optional - advanced rate limiting)
# Flask-Limiter==3.3.1
# redis==4.6.0

# Optional: For advanced NLP (if needed later)
# transformers==4.21.0
# torch==2.0.1